Modulo_4/CAP08_CHALLENGE/chatbot/requirements.txt
python-dotenv
openai
serper
requests